requirements.txt
PyQt6>=6.4.0
requests>=2.28.0
python-telegram-bot>=20.0
